Bug fixes and cleaned up the dependency_injection_test
