+++ announcement_service.py
@@ -5,7 +5,7 @@ class AnnouncementService(BotCommandServiceBase):
         super().__init__(config, "announcement_service", discord_service)
 
     async def bot_command_callback(self, message):
-        msg_tokens = message.split(' ')
+        msg_tokens = message.content.split(' ')
         channel = msg_tokens[1]
         announcement = ' '.join(msg_tokens[2:])
