+++ dependency_injection.py
@@ -7,6 +7,7 @@ from welcome_message import WelcomeMessage
 from user_roles_service import UserRolesService
 from cached_configuration_service import CachedConfigurationService
 from user_reaction_report.user_reaction_report import UserReactionReport
+from announcement_service import AnnouncementService
 
 class Dependencies:
     def __init__(self, config_path):
@@ -19,6 +20,7 @@ class Dependencies:
         self.user_leave_notification = providers.Singleton(UserLeaveNotification, self.config, self.discord_service, self.discord_mention_factory)
         self.user_roles_service = providers.Singleton(UserRolesService, self.config, self.discord_service)
         self.user_reaction_report = providers.Singleton(UserReactionReport, self.discord_service, self.config)
+        self.announcement_service = providers.Singleton(AnnouncementService, self.config, self.discord_service)
 
     def _create_config_service(self, config_path):
         return providers.Singleton(CachedConfigurationService, config_path)
